WIP: allowing MenuBar to determine its location
